Improve db restore command
Allow to "stun" database just after restore (useful for testing)

Allow to restore db in "selfish" mode, that stops server (if it is
running) before restoring database, and starts it again after database
restored. This is useful to ensure that no other odoo process touch
database while restoration of database is in progress.